refactor: stream을 사용해 GameResult 취합
- collect group by identity and counting
- counting() 결과에따라 GameResult 내부 컬렉션의 밸류 타입 변경

+++ src/main/java/edu/nextstep/camp/lotto/view/OutputView.java
@@ -17,7 +17,7 @@ public class OutputView {
         }
     }
 
-    public static void printGameResult(Map<Rank, Integer> ranks, long totalPrize, float earningRate) {
+    public static void printGameResult(Map<Rank, Long> ranks, long totalPrize, float earningRate) {
         System.out.println("당첨 통계");
         System.out.println("---------");
         System.out.println("3개 일치 (" + Prize.FOURTH_PRIZE + "원)-" + ranks.get(Rank.FOURTH));
